Colocate route-specific server actions with their routes
- Move createComment() to app/[post]/actions.ts (used only by /[post])
- Move getUserPosts() to app/userposts/actions.ts (used only by /userposts)
- Keep createPost() and deletePost() in app/actions.ts (createPost on /, deletePost shared)
- Update component imports to use absolute paths
- Follows Next.js best practice: colocate files near their usage

This improves code organization and clarity per the local-nextjs-routing-structure guidance.
